fix: correct parameter type in PageTimings.setOnContentLoad()
Changed setOnContentLoad() parameter from int to float to match the
property type and getter return type. Updated test to use float values
(1234.56 and 789.12) to validate the fix works correctly.

This resolves a type mismatch that would fail with strict types enabled.

+++ src/PageTimings.php
@@ -32,7 +32,7 @@ final class PageTimings
         return $this->onContentLoad;
     }
 
-    public function setOnContentLoad(int $onContentLoad): self
+    public function setOnContentLoad(float $onContentLoad): self
     {
         $this->onContentLoad = $onContentLoad;
